% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SL.twostg.R
\name{make_template}
\alias{make_template}
\title{Helper function for creating 2 stage libraries}
\usage{
make_template()
}
\value{
character string to be used as function
}
\description{
SL.twostg uses the string returned by this function as a template for creating two stage learner functions.
Gsub is used to substitute the names of the desired learners in for "STG1_FN" and "STG2_FN".
}
