% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SL.twostg.R
\name{predict.SL.twostg}
\alias{predict.SL.twostg}
\title{Prediction for SL.twostg}
\usage{
\method{predict}{SL.twostg}(object, newdata, ...)
}
\arguments{
\item{object}{SL.twostg object}

\item{newdata}{Dataframe to generate predictions}

\item{...}{unused additional arguments}
}
\description{
Prediction for SL.twostg
}
