% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SL.twostg.R
\name{SL.twostg}
\alias{SL.twostg}
\title{Create two stage functions}
\usage{
SL.twostg(
  stg1_fn = c("SL.glm", "SL.ranger"),
  stg2_fn = c("SL.glm"),
  twostg_env = .GlobalEnv
)
}
\arguments{
\item{stg1_fn}{list of stage 1 function(s)}

\item{stg2_fn}{list of stage 2 function(s)}

\item{environment}{to create functions within, default to global environment}
}
\value{
list of newly created library names in format "SL.stg1.STG1NAME__stg2.STG2NAME"
}
\description{
Used to create two stage functions for SuperLearner
}
