% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SL.twostg.R
\name{SL.twostg}
\alias{SL.twostg}
\title{Create two stage functions}
\usage{
SL.twostg(
  stg1_fn = c("SL.glm", "SL.ranger"),
  stg2_fn = c("SL.glm"),
  twostg_env = .GlobalEnv
)
}
\arguments{
\item{stg1_fn}{list of stage 1 function(s) (binomial family)}

\item{stg2_fn}{list of stage 2 function(s) (gaussian family)}

\item{environment}{to create functions within, default to global environment}
}
\value{
list of newly created library names in format "SL.stg1.STG1NAME__stg2.STG2NAME"
}
\description{
Used to create two stage functions for SuperLearner
}
\examples{
#simulate data
Y <- rnorm(100)
X <- data.frame(X=rnorm(100), Z=rnorm(100))

# create two stage libraries for SL.glm (1) and SL.ranger (2), SL.glm (1) and SL.glm (2)
twostg_libnames <- SL.twostg(stg1_fn = c("SL.glm", "SL.ranger"), stg2_fn = c("SL.glm"), twostg_env = .GlobalEnv)

#fit SuperLearner using newly created two-stage libraries
fit <- SuperLearner(Y=Y, X=X, SL.library = twostg_libnames)

# get back predictions on new data
newX <- data.frame(X=rnorm(100), Z=rnorm(100))
pred_twostg <- predict(fit, newX)

}
